Added to script.js, Added to mail.php
Functions:
-checkValidation
Validates form inputs and if wrong, activates the invalid fields to notify that the user needs to change the input. The invalid fields will display the proper text if either the fields are empty, or the pattern does not match.

-writeModal
This function writes the form input to the modal for the user to check one last time that the information is correct before sending it off.

-successMessage
Shows success message on the modal after form has been sent.

-closeModal
This function closes the modal


mail.php:
Gathered information from the form with the $_POST method and adds it to variables before sending the information with the mail() function.